Move BRNNs to one module
